chore(dependencies): Upgrade Spring Boot to 2.2.1 (#3333)
* fix(templated-pipelines-v1): Expected artifacts correctly inherited/merged

TemplateMerge.mergeDistinct() was merging 2 list of different types into one. Upcoming Jackson library update  would have broken compilation of TemplatedPipelineModelMutator if left unmodified.

* chore(dependencies): Upgrade Spring Boot to 2.2.1

Adjusting code for Jedis 3, Jackson 2.10 upgrades.

* fix(artifacts): Always persisting `pipeline.expectedArtifacts.boundArtifact`

Artifact resolution wasn't always persisting `pipeline.expectedArtifacts.boundArtifact` depending on the expectedArtifact's type when passed to `ArtifactResolver.resolveArtifacts()`.
Only `ExpectedArtifact` would get their `boundArtifact` updated; if a Map was used then a copy of that Map would be updated instead and this isn't returned to the caller.
The update of Jackson to 2.10 made it so that `boundArtifact` would always be updated on a copy of the passed `expectedArtifact` no matter its type, this commit is an attempt to fix both problems.
A better fix would be to make it clear what type is accepted and expected, but that's a much larger refactor.

* chore(dependencies): Upgrade Spring Boot to 2.2.1 cont'd

* jackson-module-kotlin changed the serialization behaviour for fields starting with 'is': https://github.com/FasterXML/jackson-module-kotlin/issues/80
* jackson 2.10's objectMapper.convertValue() now converts everything even if the source is already from the target type. The fixed test was expected a value from a not fully converted Map.

* chore(dependencies): bump kork and keiko