fix(kubernetes): Fix stability for stateful/daemon set (#4239)
* test(kubernetes): Add tests to stateful/daemon set stability

The next commit will fix bugs in the stateful set and daemon
set stability conditions. Add tests to cover the cases that
will be fixed; in the case of the daemon set handler, there
were no existing tests so this commit adds some basic tests
in addition to tests of the bug that will be fixed.

This commit also does some minor cleanup on the stateful set
tests, removing some unused variables and referring to member
variables by their name instead of via their groovy-generated
accessors.

* fix(kubernetes): Fix stability for stateful/daemon set

When scaling a stateful set down to zero replicas, we never
consider the stateful set stable. This is because the status
field for a stateful set with no replicas omits some fields
(instead of setting them to 0) and our logic does not properly
handle the null case. Fix this and add some tests to validate
the fix.

A similar bug exists with daemon sets; for daemon sets, you
don't explicitly set replicas but if you configure the daemon
set such that it runs on 0 nodes (for example by including a node
selector that doesn't match anything), Spinnaker will never
consider the daemon set stable.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>