fix(kubernetes): Delete namespace caching agent (#4246)
The namespace caching agent does not correctly handle the case
where namespaces are not readable, and continuously throws
errors. The KubernetesCoreCachingAgent does correctly handle
kinds that are not readable; there is no reason for namespaces
to be special-cased like this.  The KubernetesCoreCachingAgent
already enumerates all known kinds (both namespaced and
cluster-scoped) and caches them.  (It's actually already caching
namespaces as there is no special-case logic to exclude them.)

Just delete the KubernetesNamespaceCachingAgent and allow
namespaces to be cached by the KubernetesCoreCachingAgent.