refactor(plugins): makes SpinnakerPluginDescriptor a subclass of DefaultPluginDescriptor. (#437)
This was largely to implement equals/hashCode but also plays slightly nicer with the
expected extension mechanism of PF4J.

There is a slightly worse immutability story here, but the base PluginDescriptor is
already a fully nullable POJO so I don't think it is too bad to have our new
attributes as var as well.

Also adds some PluginLoading tests (the original driver for the desire to have an
equality test on SpinnakerPluginDescriptor..). These are implemented as a generic
TCK as I have a follow-up PR that introduces a new plugin-type that exercises the
same set of tests for that type.