fix(artifacts): Replace implicit use of all-args Artifact constructor (#4259)
A few tests are implicitly using the all-arg Artifact constructor,
which is being made private. These tests were relying on groovy's
automatic return value casting to cast a mock return value from a
map to an artifact. The automatic casing was relying on the all-args
constructor, which has been removed.  The fix is to construct the
artifact directly using the builder.